T442: splitting to files and packaging
